device list and configuration files are hardcoded
Initially, the location of the config file was an argument to both of
the scripts. However, now that the program is going to be deployed in an
entirely automated fashion, rather than interactive, it is unlikely that
we will need the flexibility of having the choice of config files.
Any changes to the software's configuration can be implemented through
releases or patches.

In addition the location of the devices.json file that defines the relationship
and properties of the manufacturers and their devices in the study is also
hardcoded into both scripts.
In future, this information will likely be stored in a database rather
than a file.

Finally, minor changes have been made to the unit tests, so that there
is a test_<x>.py file for each module when necessary, rather than
grouping tests by shared functionality.