Release/0.2.1 (#48)
* Scraping limits are dates rather than datetimes (#47)

Previously datetimes were used which added an unncessary complication to
any time calculations. This caused particular problems when generating
the datetimes to filter data in the API queries, as some manufacturers
use inclusive limits and some exclusive.
AQMesh and Zephyr were losing data recorded after 23:58 on a given day
as a result.

Now, the program uses _inclusive_ dates to define the scraping
timeframe, with the conversion to datetimes being handled by each
Manufacturer.scrape_device method. This makes it more obvious that is is
the responsibility of the Manufacturer sub-class to convert an inclusive
date into whatever format is required for the API call.

* Preparing for release