Scraper runs on Fargate
A large amount has been changed to get the scraper running on a
container in Fargate, as opposed to a container locally.

The scraper now has an entry script, `entry.py`, which calls the
scraping executable, the pre-processing executable, and will call a
script to send an email.
This entry point allows for arguments to specify aspects of the
scraping, such as which devices to include and which date to scrape for.
These values can be overwritten in Fargate when creating a scheduled
task.
This is a cleaner setup than previously, which used 4 `run_<x>.py`
scripts to run tasks in the locally stored container, and has increased
flexibility.

A Makefile has been provided that builds, tags, and pushes the latest release
to ECR. It can automatically scrape the version number using the value
provided in `setup.py`, through the new `version.sh` script.
It relies upon a file called `deploy.env` containing repository properties to
be available in the working directory. See the Makefile for further
details.