Added a launcher script to run Fargate tasks
The new `run_scrape.py` script allows the user to run single scraping
tasks on Fargate. This is the easiest way to run ad-hoc one-off tasks
without having to use the aws-cli, as you cannot start tasks from the
AWS web interface.

It has arguments to allow the user to specify which devices to scrape
for and on what day, and whether to upload the resulting data to Google
Drive or not.

It relies upon having a `run.env` file in the working directory that
contains AWS runtime properties. See the docstrings inside
`run_scrape.py` for further details.