Moved google credentials from JSON to env var
Previously the Google Service account was authenticating by providing a
JSON file. However, this meant that the JSON file had to be embedded
into the Docker image for deployment.

Now, the credentials are loaded as environment variables, with the
GOOGLE_CREDS env var expecting a JSON-parseable string containing the
relevant properties.