Feature/fargate deploy (#50)
* Refactor to facilitate AWS deployment

Several large changes to how the CLI scraping program is run have been
introduced in order to faciltate deploying the scraper on AWS:

  - Credentials are passed in through environment variables, as required
    by AWS. To help development, a .env file (not included in repo) has
    been created holding the credentials, which is loaded using the
    `dotenv` package. The variables found in the .env file are only
    loaded into the environment if they are not already present (as
    would be the case in production on AWS). This allows for the same
    configuration to be used in development and production.
  - A new argument `--devices` has been created, which allows for the
    specification of which devices should be scraped, defaulting to all
    available devices if argument isn't provided.
  - To provide a reference of all the available devices and the
    manufacturers they belong to, as required by the new `--devices`
    flag, a JSON file (devices.json, not included in repo) defines the
    relationship between manufacturers and the devices they own. This
    was previously achieved in the .ini file, but expressing
    hierarchical relationships in a flat format proved challenging.
    Ideally these relationships would be obtained from a database.
  - To reflect the data entity relationship between
    Manufacturers and Devices in devices.json, made explicit in
    devices.json, a Device class has been added. Manufacturers now have
    a reference to every Device they own, and Devices own their own
    data.

* Moved scraping window dates to scrape_device from constructor

* Refactored how summary tables are stored

* device list and configuration files are hardcoded

Initially, the location of the config file was an argument to both of
the scripts. However, now that the program is going to be deployed in an
entirely automated fashion, rather than interactive, it is unlikely that
we will need the flexibility of having the choice of config files.
Any changes to the software's configuration can be implemented through
releases or patches.

In addition the location of the devices.json file that defines the relationship
and properties of the manufacturers and their devices in the study is also
hardcoded into both scripts.
In future, this information will likely be stored in a database rather
than a file.

Finally, minor changes have been made to the unit tests, so that there
is a test_<x>.py file for each module when necessary, rather than
grouping tests by shared functionality.

* Added CLI arguments

Several runtime configuration values, such as scraping windows, which
devices to include, whether to upload the results to Google Drive etc...
have been moved from flags in `.ini` files to command line arguments.
Both `quant_scrape` and `quant_preprocess` have had these changes.

These changes are to make it easier to run the programs from AWS.
Currently, the scraping is run manually so it is easy to quickly modify
a config file before running the program.
However, on AWS it is hard to modify files and then pass them into
programs. Instead, it is much easier to pass in options through program
arguments.

Additional changes are minor stylistic improvements suggested by pylint.

* Minor refactor and improved error handling

I refactored the loading devices.json functionality into a separate
function in utils and added error handling and unit tests.
I also added informative logging warnings to the factory that
instantiates Manufacturers and Devices from the devices.json config.
I previously wasn't keen on adding logging calls into a library
function, but I can't see how else to do it

* Scraper runs on Fargate

A large amount has been changed to get the scraper running on a
container in Fargate, as opposed to a container locally.

The scraper now has an entry script, `entry.py`, which calls the
scraping executable, the pre-processing executable, and will call a
script to send an email.
This entry point allows for arguments to specify aspects of the
scraping, such as which devices to include and which date to scrape for.
These values can be overwritten in Fargate when creating a scheduled
task.
This is a cleaner setup than previously, which used 4 `run_<x>.py`
scripts to run tasks in the locally stored container, and has increased
flexibility.

A Makefile has been provided that builds, tags, and pushes the latest release
to ECR. It can automatically scrape the version number using the value
provided in `setup.py`, through the new `version.sh` script.
It relies upon a file called `deploy.env` containing repository properties to
be available in the working directory. See the Makefile for further
details.

* Added a launcher script to run Fargate tasks

The new `run_scrape.py` script allows the user to run single scraping
tasks on Fargate. This is the easiest way to run ad-hoc one-off tasks
without having to use the aws-cli, as you cannot start tasks from the
AWS web interface.

It has arguments to allow the user to specify which devices to scrape
for and on what day, and whether to upload the resulting data to Google
Drive or not.

It relies upon having a `run.env` file in the working directory that
contains AWS runtime properties. See the docstrings inside
`run_scrape.py` for further details.

* Removed Google Drive folder IDs from .ini files

Google Drive folder IDs have instead been placed in environment
variables, so they can be more securely passed into the application when
running in a container, rather than having to bake the .ini file into
the container.

* Moved google credentials from JSON to env var

Previously the Google Service account was authenticating by providing a
JSON file. However, this meant that the JSON file had to be embedded
into the Docker image for deployment.

Now, the credentials are loaded as environment variables, with the
GOOGLE_CREDS env var expecting a JSON-parseable string containing the
relevant properties.

* Added a script to send summary emails using SES