Added flexibility to HTML summary table cell color
Previously, cells were coloured either red, orange, or green, depending
if they had data availability of 0%, 100% or inbetween.

To allow the user to format this table as desired, the code now allows
the user to specify buckets by the lowest inclusive value for that
bucket and the corresponding cell colour.
The code cascades through the buckets in descending order until a match
is met and that cell colour is applied.
This allows for the user to have a much more granular cell colouring
scheme with far more flexibility in both the colours and the ways in
which they are applied.