Fixed bug where all raw fields were in analysis files (#68)
Previously the `daily_preprocessing.py` script limited the cleaned data
files to only gases and PM measurements of interest. However, in the
last update this got reverted and all the columns in the raw data were
also included. The column level flag that controlled this behaviour was
stored in the old `preprocessing.ini` configuration file and got
overlooked during the transition to the new `devices.json` file.