Release/0.4.0 (#70)
* Uploads data availability as CSV to Google Drive (#59)

* Fixed bug preventing quantaq from installing (#60)

The URL to download a specific tagged release of the `quantaq` package had
started to fail, seemingly due to the `master` branch being updated for
the first time in months and the URL tag syntax not correctly resolving to
the actual release.

* Fixed dependency issue in requirements.txt (#62)

The `quantaq` dependency bug that was fixed in `setup.py` also needed
fixing in `requirements.txt`. `requirements.txt` is only used by
GitHub Actions to install the package dependencies, where `setup.py` is
used to install the package itself (including its dependencies) into a
production environment.

* Feature/emails dkim (#63)

* send_email.py uses an Identity ARN for authorisation

* LogGroup created in CloudFormation

Previously this was manually created. It is now setup automatically to
reduce the amount of manual configuration required when installing the
stack.

* Updated setup docs

* Only one date displayed in summary email

Previously the summary email displayed the text "QUANT scraping summary
from <x> to <y>". However, given that nearly every use case of this
scraping has been to scrape a single day's worth of data, the upper
limit was redundant and has been removed.

* Feature/cloudformation scheduledtask (#65)

* Fixes bug in previous commit

* ScheduledTask has default parameters

ScheduledTask now has default parameters for uploading all data to
Google Drive and sending the summary email to the account specified
during the Stack creation.

* Feature/send email entrypoint (#66)

* Fixes bug in previous commit

* Made send_email an entry point.

Helps to be consistent as this is a module level script along with
cli.py and daily_preprocessing.py yet it was treated differently before.
Also have moved example config files into `resources` to tidy up top
level directory.

* Minor logging update

* Added tags to resources

* Fixed bug where all raw fields were in analysis files (#68)

Previously the `daily_preprocessing.py` script limited the cleaned data
files to only gases and PM measurements of interest. However, in the
last update this got reverted and all the columns in the raw data were
also included. The column level flag that controlled this behaviour was
stored in the old `preprocessing.ini` configuration file and got
overlooked during the transition to the new `devices.json` file.

* Feature/scrape defra (#69)

* Working scraping of Fishergate AURN site

The scraping works as part of the regular routine and just needs some
unit tests adding.
It isn't an ideal solution, however, as the AURN API doesn't fall under
the clear Manufacturer/Device/Pollutant hierarchy that the other
webportals use.
The AURN API gives each pollutant a unique ID rather than each
recording site (the analogue of a device). So scraping 3 pollutants from
one site is treated the same as scraping 1 pollutant from 3 different sites.

The current solution won't work well if we expand to multiple AURN
sites. One solution would be to have Device properties available in
the scrape_device function and then store the pollutant IDs per device
rather than per manufactuer in devices.json

* Improved error handling and unit tests

* AQMesh changed login request structure

Previously credentials were passed in using the standard urlencoded
manner, but it's been changed around the 10th July to use multipart
instead.

* Data availability CSV has pct avail columns

* Bumped version number