Added application configuration to git
The files devices.json, config.ini, and preprocessing.ini weren't
previously saved in version control. However, there is no need for them
not to be, as the most sensitive information contained in them are the
device IDs. However, there is no security risk from having these IDs in
a private repo.