Release/0.5.0 (#72)
* Feature/aws deploy (#71)

* Added application configuration to git

The files devices.json, config.ini, and preprocessing.ini weren't
previously saved in version control. However, there is no need for them
not to be, as the most sensitive information contained in them are the
device IDs. However, there is no security risk from having these IDs in
a private repo.

* Scheduled Task fully described in CloudFormation.

Previously, device IDs weren't included in CloudFormation, however it
has been confirmed that they aren't sensitive information and can thus
be included in the CF template.

* Removed unused build file and setting up FIXME precommit

* Added FIXME checker to precommit

* Initial test of deploy github action

* test 2

* Fixed incorrect filename

* AWS user ID stored in secret

* Tag ECR image with latest

* Improved docs

* Removed unused build file and setting up FIXME precommit

* Action only runs when PR is closed.

By default, an Action triggers a "pull-request" when the PR is opened.
The intended behaviour for a release action instead should by triggered
by the PR being closed.

* Using updated deploy action

The University have released a new version of the CF deploy action,
which correctly applies tags to the stack.

* Bumped version number