Feature/aqmesh api (#74)
* AQMesh uses API

Previously AQMesh data was downloaded by interfacing with HTTP requests
with the airmonitors web-portal. It now uses the API which should be
quicker, more robust, and more secure as we don't need to store a
password

* Cleaned unit tests

Previously a success criterion of unit tests was whether an exception or
not was raised. However, if the output is compared to an expected value
then this test for an exception is redundant as the output will not
returned