Fixed permissions bug from last release
The ecs:RunTask permission on the scheduled deployment role is limited
to the scraping Task item. Previously this was denoted in CF as
"QUANTTasks:*", to denote all releases from this task. However, when the
stack was automatically deployed in release 0.5.0, this syntax no longer
worked. Instead, the syntax "QUANTTasks" is needed.