Working implementation
The core functionality is implemented. This involves identifying which
PA files are available in the raw source folder but not in the Clean
QUANT repository, downloading them, validating them and formatting them
into the standard shape, and uploading them to the Clean folder.
A summary email can also be sent like with the main scraping routine.

Outstanding tasks include adding the Analysis preprocessing script, as
well as refactoring a large amount of duplicate code with the main
CLI script.