Release/0.6.0 (#83)
* Feature/aqmesh api (#74)

* AQMesh uses API

Previously AQMesh data was downloaded by interfacing with HTTP requests
with the airmonitors web-portal. It now uses the API which should be
quicker, more robust, and more secure as we don't need to store a
password

* Cleaned unit tests

Previously a success criterion of unit tests was whether an exception or
not was raised. However, if the output is compared to an expected value
then this test for an exception is redundant as the output will not
returned

* Fixed permissions bug from last release

The ecs:RunTask permission on the scheduled deployment role is limited
to the scraping Task item. Previously this was denoted in CF as
"QUANTTasks:*", to denote all releases from this task. However, when the
stack was automatically deployed in release 0.5.0, this syntax no longer
worked. Instead, the syntax "QUANTTasks" is needed.

* Build tests only run when PRs are opened

Previously this action was also run on pushes, which resulted in
duplication when pushing to a branch that then had a PR opened.

* Syntax fix

* Feature/purpleair (#79)

* Moved Docker file into separate folder

Will need 2 Dockerfiles, one for the scraper and one for the purple air
conversion script. This commit puts the scaffolding in place to have
multiple Dockerfiles

* Working implementation

The core functionality is implemented. This involves identifying which
PA files are available in the raw source folder but not in the Clean
QUANT repository, downloading them, validating them and formatting them
into the standard shape, and uploading them to the Clean folder.
A summary email can also be sent like with the main scraping routine.

Outstanding tasks include adding the Analysis preprocessing script, as
well as refactoring a large amount of duplicate code with the main
CLI script.

* Minor refactoring

* Added preprocessing

A second script has been added that identifies which dates do not
currently have 'Analysis' PurpleAir files and sets up the call to
`quant_preprocess` to generate them.

* Minor refactoring

The date format used in QUANT filenames, and the timestamp format used
in cleaned files are now configurable options in `config.ini`.

* Added uploading PurpleAir availability

* Added unit tests

* PurpleAir conversion script runs on Fargate

The CloudFormation template has also been refactored by separating out
the scheduled scraping task arguments (which devices to use etc...) and
placing them in `cloudformation/configs/<production/dev>.json`

* Set Google Drive logging to warning level

By default every request to the Drive API is logged, which pollutes the
log particularly when making lots of calls to files.list().

* Deleted files are not included when listing files

The default undocumented behaviour of the Google Drive API is to include
files in Shared Folders that have been deleted by one user.
`list_files_googledrive()` now explicitly doesn't include these deleted
files by default, although this behaviour can be overridden.

* Fixed timestamp format bug

* Improved logging

* Replaced deprecated ADD with COPY

* Release/0.5.1 (#75)

* Updated webid for Zep344

This devices has been repaired but has been assigned a new ID by Zephyr
- 540. We will continue to refer to it as 344 since the sensors haven't
  been modified, just the comms.

* Fixed permission issue

Previously the permission 'ecs::RunTask' was limited to the scraping
task and all future versions by specifing `<taskname>:*` in the policy.
However, this syntax no longer works and the `:*` is now unnecessary.

* Bumped version number

* Renamed master branch to main (#77)

* AQMesh changed login request (#78)

AQMesh have made a minor change to their login request, so that the
username field has been renamed to 'emailaddress'.

* Feature/aqmesh api (#74)

* AQMesh uses API

Previously AQMesh data was downloaded by interfacing with HTTP requests
with the airmonitors web-portal. It now uses the API which should be
quicker, more robust, and more secure as we don't need to store a
password

* Cleaned unit tests

Previously a success criterion of unit tests was whether an exception or
not was raised. However, if the output is compared to an expected value
then this test for an exception is redundant as the output will not
returned

* Bumped version number

* Bugfix preventing deployment

* Fixed bug stopping environment vars loading

This is caused by moving the email sending code inside the entry point
scripts, rather than as a separate script. These emails need secrets
setting, which wasn't being done.