Bugfix for PurpleAir uneven number of fields (#92)
CSVs recorded on PA devices often have a different number of columns in
each row. It appears that if a field isn't available, the comma
delimiter isn't added to the CSV output. This results in an unparseable
CSV, i.e. if you expect 10 fields but only get 8, how do you know which
8 these values correspond to?

The solution I've taken is to assume the number of headers will always
be correct and remove any rows that don't have this many fields