Release/0.7.0 (#94)
* Bump release number

* Feature/aeroqual api (#93)

* Working Aeroqual API implementation

NB: Need to get unit tests working and add error handling for
Aeroqual.parse_to_csv

* Moved TODOs to Issues

* Added unit tests for scrape_data and parse_to_csv

* Added scraping instrument calibration parameters

There isn't a way to access these from the API so I've added in the old
web page scraping function to obtain these values.

* Bugfix for PurpleAir uneven number of fields (#92)

CSVs recorded on PA devices often have a different number of columns in
each row. It appears that if a field isn't available, the comma
delimiter isn't added to the CSV output. This results in an unparseable
CSV, i.e. if you expect 10 fields but only get 8, how do you know which
8 these values correspond to?

The solution I've taken is to assume the number of headers will always
be correct and remove any rows that don't have this many fields