Fixed bug with parsing Aeroqual timestamps (#95)
The code was raising an uncaught error when attempting to parse
erroneous timestamps in Aeroqual data. This is because the parsed
Aeroqual data now comes in as float from the API, as opposed to string
from the old website scrape.
If a float value is present in the timestamp column, then the
`datetime.strptime` call raises a `TypeError`, rather than a
`ValueError` which occurs if the input is a string.
The code now catches `TypeErrors`.