Parse failures are indicated in summary email
Previously, if the CSV parsing failed for a specific recording, this was
omitted from the summary email and availability statistics.
This has been corrected, so that the only situation where a device will
not be present in a given day's summary is if it did not have a raw
file for this date.