Reauthenticate Google API between uses
We were starting to observe errors when uploading files to Google Drive,
with one possible cause being that the same connection handle was being
used throughout the entire runtime of the program.
The errors typically occurred when attempting to upload QuantAQ files,
whose scraping takes the longest.
One possible case then could be that Google have started to terminate
 unused connections, and since this scraping of QuantAQ takes a long
 time, it is identified as idle.

 I've changed the code to reconnect between each batch of uploads in an
 attempt to mitigate this problem.