Working scraper
The `MyQuantAQ` class has been extended to support the Modulair devices
that are accessed through the same QuantAQ API. The only real difference
is that met variables are stored in a nested object within a record
object for a particular timepoint. Our previous devices only had lat/lon
as nested objects and this was hardcoded into the `parse_to_csv` method.
This has now been refactored so that any nested objects get flattened.

The Modulair devices have been added as belonging to the QuantAQ
manufacturer, but it might be decided to keep them as a separate entity.
If so, this will simply require subclassing `MyQuantAQ` and adding a new
entry in `devices.json` and a corresponding reference in
`quantscraper/factories.py`.