Added EnvironemntalInstruments subclass
EI uses the same AirMonitors API as the AQMesh devices we already
scrape. To differentiate between devices made by these 2 different
companies, a sub-class of AQMesh has been setup for EI.

Currently it can't be assessed as the API isn't returning data for our
devices - confirmed to be a server side problem.