Release/expanded study (#121)
* Scaffolding for second scheduled scrape

Sebastian has requested that the wider participation devices be scraped
separately to the initial study and summarised in a separate email.
This is simply facilitated by creating a second `AWS::Events::Rule`.

* Fixed bug with PA emails

* Parameterised email subject

* Parameterised upload destinations as container arg

Previously, the folders where the resultant cleaned data were to be
uploaded to were specified as env vars.
This commit changes them to be input into the container as command-line
arguments, providing a bit more transparency and making it easier for
different runs to upload to different locations.

* Feature/south coast science (#112)

* Working scraper

The scraper can download data for SCS devices. Outstanding features
include unit tests, and specifying which fields to include following
discussion with project collaborators.

* Added PM to scraper

* Minor unit test changes

* Feature/clarity (#113)

* Scaffolding

* Working scraper

The scraper is working. Outstanding tasks before this feature can be
finalised are confirming with the researchers which fields need to be
scraped, and adding unit tests

* Added our third device and cleaned device names

* Feature/vortex (#114)

* Working scraping routine

The only outstanding task is to add unit tests.

* Renamed measurement fields + applied scaling

The fields in the returned data have changed labels slightly.
The gases are scaled to ppb now.

* Feature/quantaq wider participation (#115)

* Working scraper

The `MyQuantAQ` class has been extended to support the Modulair devices
that are accessed through the same QuantAQ API. The only real difference
is that met variables are stored in a nested object within a record
object for a particular timepoint. Our previous devices only had lat/lon
as nested objects and this was hardcoded into the `parse_to_csv` method.
This has now been refactored so that any nested objects get flattened.

The Modulair devices have been added as belonging to the QuantAQ
manufacturer, but it might be decided to keep them as a separate entity.
If so, this will simply require subclassing `MyQuantAQ` and adding a new
entry in `devices.json` and a corresponding reference in
`quantscraper/factories.py`.

* Added Modulair as a separate class

* Feature/kunak (#116)

* Working scraper

* Removed unused raw PM fields

* Working scraper (#117)

* Feature/bosch (#118)

* Working scraper

Next steps are to add unit tests and convert NO2 and O3 to ppb

* Converted ug/m3 to ppb

* Feature/respirer (#119)

* Working scraper

Final steps before release are to confirm with the manufacturer and the
researchers which fields we need to include in the Clean data, confirm
the final Manufacturer name, and add unit tests.

* Removed unused fields from clean data

* Feature/environmental instruments (#120)

* Added EnvironemntalInstruments subclass

EI uses the same AirMonitors API as the AQMesh devices we already
scrape. To differentiate between devices made by these 2 different
companies, a sub-class of AQMesh has been setup for EI.

Currently it can't be assessed as the API isn't returning data for our
devices - confirmed to be a server side problem.

* Added CO field

* Fixing merge issues

* Renamed devices/manufacturers

We have agreed upon a standard set of labels for the companies and their
devices in this study

* Add SCS VOC to clean data

* Deployment prep

* Modular files are prefixed Modulair

* Prep for release

* Renamed SCS Ox field to O3

SCS have advised us that they use "Ox" interchangeably with "O3" to
refer to Ozone measurements. Ref email with Ivan on 2021-06-29

* Bosch gases changed to ppb

* Tweaked SCS column names

The exegesis calibration products are now the primary measurement for
their respective pollutant following a call with the SCS team.

* Bump release number