Fixed rate limit issue
The Oizom API allows a maximum of 4 requests a second.
This lead to errors when scraping more than one device at a time, as for
each device there are 3 API calls (2 to get status, 1 to get data).
To fix this, a 1-second pause has been added at the end of
log_device_status() and scrape_device().