Fixed bug introduced by API change (#155)
The returned data from QuantAQ now also contains model parameters. This
adds a third level of hierarchy to the JSON, which the current code
cannot handle. This fix simply removes the problematic columns, but
doesn't solve the underlying issue of having 3D data. In future it might
be preferable to work with the raw JSON data rather than trying to
coerce it into a flat DataFrame asap.